refactor(shared): poktesting package
