+++ app/client/keybase/debug/keystore.go
@@ -6,10 +6,9 @@ import (
 	"sync"
 
 	"github.com/pokt-network/pocket/app/client/keybase"
-	"github.com/pokt-network/pocket/build"
 	"github.com/pokt-network/pocket/logger"
 	cryptoPocket "github.com/pokt-network/pocket/shared/crypto"
-	"gopkg.in/yaml.v2"
+	"github.com/pokt-network/pocket/shared/poktesting"
 )
 
 const (
@@ -43,7 +42,7 @@ func initializeDebugKeybase() error {
 		err                  error
 	)
 
-	validatorKeysPairMap, err = ParseValidatorPrivateKeysFromEmbeddedYaml()
+	validatorKeysPairMap, err = poktesting.ParseValidatorPrivateKeysFromEmbeddedYaml()
 
 	if err != nil {
 		return err
@@ -124,26 +123,4 @@ func initializeDebugKeybase() error {
 	}
 
 	return nil
-}
-
-// ParseValidatorPrivateKeysFromEmbeddedYaml fetches the validator private keys from the embedded build/localnet/manifests/private-keys.yaml manifest file.
-func ParseValidatorPrivateKeysFromEmbeddedYaml() (map[string]string, error) {
-
-	// Parse the YAML file and load into the config struct
-	var config struct {
-		ApiVersion string            `yaml:"apiVersion"`
-		Kind       string            `yaml:"kind"`
-		MetaData   map[string]string `yaml:"metadata"`
-		Type       string            `yaml:"type"`
-		StringData map[string]string `yaml:"stringData"`
-	}
-	if err := yaml.Unmarshal(build.PrivateKeysFile, &config); err != nil {
-		return nil, err
-	}
-	validatorKeysMap := make(map[string]string)
-
-	for id, privHexString := range config.StringData {
-		validatorKeysMap[id] = privHexString
-	}
-	return validatorKeysMap, nil
 }
